aws_eks_cluster: Increase create timeout to 30 mins
It seems that as reported in #5165, an EKS cluster frequently takes more to
actually bring up than the default create timeout allows for. While the timeout
is configurable, the default should be sufficient to actually provision a
cluster in most cases.

This commit bumps the default timeout to 30 minutes.

Fixes #5165.